Add ClientWithSortKey in.
The purpose of this commit is to introduce `ClientWithSortKey`, as well
as updating the method signatures of all the `internalClient` to include
a `sortKey` parameter. These parameters are currenty just passed through
to other functions, but don't actually affect anything (other than log
messages).

I have inserted `// TODO`s throughout the code where I will be updating
the business logic in the next PR.